+++ src/main/java/com/example/rv/pojo/VehicleMaintenance.java
@@ -9,7 +9,7 @@ public class VehicleMaintenance {
   private int vehicleMaintenanceVehicleId;
   private String vehicleMaintenanceMaintenanceDetails;
   private java.sql.Timestamp vehicleMaintenanceCompletedAt;
-  private short vehicleMaintenanceStatus;
+  private int vehicleMaintenanceStatus;
   private java.sql.Timestamp vehicleMaintenanceCreatedAt;
   private java.sql.Timestamp vehicleMaintenanceUpdatedAt;
