+++ src/main/java/com/example/rv/pojo/Vehicles.java
@@ -10,10 +10,10 @@ public class Vehicles {
   private int vehicleType;
   private String vehicleLocation;
   private String vehicleDescription;
-  private float vehiclePrice;
+  private double vehiclePrice;
   private String vehiclePicture;
   private String vehicleAuditFailedMsg;
-  private short vehicleStatus;
+  private int vehicleStatus;
   private java.sql.Timestamp vehicleCreatedAt;
   private java.sql.Timestamp vehicleUpdatedAt;
